Creates images and styles folder in src directory and corrects file path locations in Tools and LandingPage components
