+++ frontend/src/App.css
@@ -1,0 +1,5 @@
+.center {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
